# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chess Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/{username}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user statistics
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
  /users/me/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /puzzles:
    get:
      operationId: listPuzzles
      tags: [puzzles]
      summary: List puzzles
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [mate1, mate2, mate3, tactical, endgame]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Puzzles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Puzzle"
  /puzzles/{id}/solve:
    post:
      operationId: solvePuzzle
      tags: [puzzles]
      summary: Submit puzzle solution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PuzzleSolveInput"
      responses:
        "200":
          description: Puzzle result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PuzzleResult"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [users]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, isGuest, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        isGuest:
          type: boolean
        avatar:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [id, username, isGuest, createdAt, ratings]
      properties:
        id:
          type: integer
        username:
          type: string
        isGuest:
          type: boolean
        avatar:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        ratings:
          $ref: "#/components/schemas/Ratings"
        recentGames:
          type: array
          items:
            $ref: "#/components/schemas/GameSummary"
    ProfileUpdate:
      type: object
      properties:
        bio:
          type: string
        country:
          type: string
        avatar:
          type: string
    Ratings:
      type: object
      required: [bullet, blitz, rapid, classical]
      properties:
        bullet:
          type: integer
        blitz:
          type: integer
        rapid:
          type: integer
        classical:
          type: integer
    UserStats:
      type: object
      required: [totalGames, wins, losses, draws, winRate, accuracy, bestRating, currentStreak, puzzleRating]
      properties:
        totalGames:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        winRate:
          type: number
        accuracy:
          type: number
        bestRating:
          type: integer
        currentStreak:
          type: integer
        puzzleRating:
          type: integer
        recentAccuracies:
          type: array
          items:
            type: number
...
    GameSummary:
      type: object
      required: [id, status, timeControl, mode, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
        result:
          type: ["string", "null"]
        resultReason:
          type: ["string", "null"]
        timeControl:
          type: string
        mode:
          type: string
        whitePlayer:
          $ref: "#/components/schemas/PlayerInfo"
        blackPlayer:
          $ref: "#/components/schemas/PlayerInfo"
        accuracy:
          type: ["number", "null"]
        opening:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PlayerInfo:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
        username:
          type: string
        avatar:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
...
    Puzzle:
      type: object
      required: [id, fen, solution, type, rating, title]
      properties:
        id:
          type: integer
        fen:
          type: string
        solution:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [mate1, mate2, mate3, tactical, endgame]
        rating:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        isDaily:
          type: boolean
    PuzzleSolveInput:
      type: object
      required: [moves, solved, timeTakenMs]
      properties:
        moves:
          type: array
          items:
            type: string
        solved:
          type: boolean
        timeTakenMs:
          type: integer
    PuzzleResult:
      type: object
      required: [solved, ratingChange, newRating, streak]
      properties:
        solved:
          type: boolean
        ratingChange:
          type: integer
        newRating:
          type: integer
        streak:
          type: integer
        correctSolution:
          type: array
          items:
            type: string
...