# lib/api-zod/src/generated/types/puzzleSolveInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chess Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PuzzleSolveInput {
  moves: string[];
  solved: boolean;
  timeTakenMs: number;
}